refactoring package layout
