+++ relay/shfu_service.go
@@ -1,0 +1,10 @@
+package relay
+
+import (
+	"github.com/datachainlab/lcp-go/relay/shfu_storage"
+)
+
+// Re-export types from storage package for backward compatibility
+type SHFURecord = shfu_storage.SHFURecord
+type SHFUHeaderRecord = shfu_storage.SHFUHeaderRecord
+type SHFUStorage = shfu_storage.SHFUStorage
