+++ relay/shfu_storage/interface.go
@@ -1,4 +1,4 @@
-package storage
+package shfu_storage
 
 import (
 	"context"
