+++ relay/shfu_storage/sqlite_dialect.go
@@ -1,4 +1,4 @@
-package storage
+package shfu_storage
 
 import (
 	"fmt"
