+++ relay/shfu_storage/sqlx_storage.go
@@ -1,4 +1,4 @@
-package storage
+package shfu_storage
 
 import (
 	"context"
