+++ relay/update_client.go
@@ -1,4 +1,4 @@
-package updater
+package relay
 
 import (
 	"context"
