--- relay/updater/service.go
@@ -1,10 +1,0 @@
-package updater
-
-import (
-	"github.com/datachainlab/lcp-go/relay/updater/storage"
-)
-
-// Re-export types from storage package for backward compatibility
-type SHFURecord = storage.SHFURecord
-type SHFUHeaderRecord = storage.SHFUHeaderRecord
-type SHFUStorage = storage.SHFUStorage
