--- relay/updater/shfu_mock.go
@@ -1,81 +1,0 @@
-package updater
-
-import (
-	"context"
-
-	"github.com/cosmos/cosmos-sdk/codec/types"
-	clienttypes "github.com/cosmos/ibc-go/v8/modules/core/02-client/types"
-	ibcexported "github.com/cosmos/ibc-go/v8/modules/core/exported"
-	"github.com/hyperledger-labs/yui-relayer/core"
-)
-
-// SHFUMockChain is a dummy implementation of core.FinalityAwareChain
-// that returns a specific height for testing SetupHeadersForUpdate calls.
-// It embeds the interface so unimplemented methods will panic at runtime.
-type SHFUMockChain struct {
-	core.FinalityAwareChain // Embedded interface - unimplemented methods will panic
-	chainID                 string
-	latestHeight            ibcexported.Height
-}
-
-var (
-	_ core.FinalityAwareChain = (*SHFUMockChain)(nil)
-)
-
-// SHFUMockClientState is a dummy implementation that embeds ibcexported.ClientState
-// All methods will panic at runtime unless specifically implemented
-type SHFUMockClientState struct {
-	ibcexported.ClientState // Embedded interface - unimplemented methods will panic
-	latestHeight            ibcexported.Height
-}
-
-var (
-	_ ibcexported.ClientState = (*SHFUMockClientState)(nil)
-)
-
-// NewSHFUMockChain creates a new SHFUMockChain instance
-func NewSHFUMockChain(chainID string, latestHeight ibcexported.Height) *SHFUMockChain {
-	return &SHFUMockChain{
-		chainID:      chainID,
-		latestHeight: latestHeight,
-	}
-}
-
-// ChainID returns the chain ID
-func (c *SHFUMockChain) ChainID() string {
-	return c.chainID
-}
-
-// LatestHeight returns the latest height with context (allowed method)
-func (c *SHFUMockChain) LatestHeight(ctx context.Context) (ibcexported.Height, error) {
-	return c.latestHeight, nil
-}
-
-// QueryClientState returns a QueryClientStateResponse with mock client state
-func (c *SHFUMockChain) QueryClientState(qctx core.QueryContext) (*clienttypes.QueryClientStateResponse, error) {
-	// Get height from query context
-	height := qctx.Height()
-	mockClientState := NewSHFUMockClientState(height)
-
-	// Pack the client state into Any type
-	clientStateAny, err := types.NewAnyWithValue(mockClientState)
-	if err != nil {
-		return nil, err
-	}
-
-	return &clienttypes.QueryClientStateResponse{
-		ClientState: clientStateAny,
-	}, nil
-}
-
-// NewSHFUMockClientState creates a new SHFUMockClientState instance
-func NewSHFUMockClientState(latestHeight ibcexported.Height) *SHFUMockClientState {
-	return &SHFUMockClientState{
-		latestHeight: latestHeight,
-	}
-}
-
-// GetLatestHeight returns the configured latest height
-func (s *SHFUMockClientState) GetLatestHeight() ibcexported.Height {
-	return s.latestHeight
-}
